refactor: improve folder structure
